backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# OCR Engines
pytesseract==0.3.10
paddlepaddle==2.6.1
paddleocr==2.7.0.3
easyocr==1.7.0
Pillow==10.1.0
opencv-python==4.6.0.66
pdf2image==1.16.3

# AI/ML
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
transformers==4.35.2
sentence-transformers>=2.2.0
huggingface_hub>=0.19.0
langchain==0.0.340
openai==1.3.7
faiss-cpu==1.12.0

# Image Processing
numpy==1.26.2
scikit-image==0.22.0

# Utilities
celery==5.3.4
redis==5.0.1
aiofiles==23.2.1
httpx==0.25.2